+++ gg-pingpong-api/src/main/java/gg/pingpong/api/user/noti/service/sns/NotiMailSender.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 
 import gg.auth.UserDto;
 import gg.data.noti.Noti;
-import gg.pingpong.api.global.utils.AsyncMailSender;
+import gg.utils.sns.AsyncMailSender;
 import gg.pingpong.api.user.noti.dto.UserNotiDto;
 import gg.pingpong.api.user.noti.service.NotiService;
 import lombok.RequiredArgsConstructor;
