+++ gg-utils/build.gradle
@@ -74,6 +74,7 @@ unitTestCoverageReport {
 
 dependencies {
     implementation "com.amazonaws:aws-java-sdk-s3:1.12.281"
+    implementation 'org.springframework.boot:spring-boot-starter-mail'
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
